Fix javadoc, and remove deprecated classes
